backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-multipart==0.0.20
httpx==0.28.1
tenacity==9.1.2
duckduckgo-search==6.3.5
pypdf==5.3.1
python-docx==1.1.2
orjson==3.11.2
